compiled all validations in one file - singleton pattern way
