--- src/auth-service/utils/validations.register.js
@@ -1,62 +1,0 @@
-const Validator = require("validator");
-const isEmpty = require("is-empty");
-module.exports = function validateRegisterInput(data) {
-  let errors = {};
-  // Convert empty fields to an empty string so we can use validator functions
-  data.userName = !isEmpty(data.userName) ? data.userName : "";
-  data.firstName = !isEmpty(data.firstName) ? data.firstName : "";
-  data.lastName = !isEmpty(data.lastName) ? data.lastName : "";
-  data.email = !isEmpty(data.email) ? data.email : "";
-  data.password = !isEmpty(data.password) ? data.password : "";
-  data.password2 = !isEmpty(data.password2) ? data.password2 : "";
-  data.privilege = !isEmpty(data.privilege) ? data.privilege : "";
-  data.organization = !isEmpty(data.organization) ? data.organization : "";
-  // Name checks
-  if (Validator.isEmpty(data.firstName)) {
-    errors.firstName = "firstName field is required";
-  }
-
-  if (Validator.isEmpty(data.lastName)) {
-    errors.lastName = "lastName field is required";
-  }
-
-  //privilege checks
-  if (Validator.isEmpty(data.privilege)) {
-    errors.privilege = "privilege field is required";
-  }
-
-  //organization checks
-  if (Validator.isEmpty(data.organization)) {
-    errors.organization = "organization field is required";
-  }
-
-  //userName checks
-  if (Validator.isEmpty(data.userName || data.email)) {
-    errors.userName = "userName field is required";
-    errors.email = "userName field is required";
-  }
-
-  // Email checks
-  if (Validator.isEmpty(data.email)) {
-    errors.email = "Email field is required";
-  } else if (!Validator.isEmail(data.email)) {
-    errors.email = "Email is invalid";
-  }
-  // Password checks
-  if (Validator.isEmpty(data.password)) {
-    errors.password = "Password field is required";
-  }
-  if (Validator.isEmpty(data.password2)) {
-    errors.password2 = "Confirm password field is required";
-  }
-  if (!Validator.isLength(data.password, { min: 6, max: 30 })) {
-    errors.password = "Password must be at least 6 characters";
-  }
-  if (!Validator.equals(data.password, data.password2)) {
-    errors.password2 = "Passwords must match";
-  }
-  return {
-    errors,
-    isValid: isEmpty(errors),
-  };
-};
